requirements.txt
langchain
Flask
streamlit
langchain_google_genai
sqlalchemy
langchain_experimental
langchain_community
langgraph
mysql-connector-python
gunicorn